requirements.txt
garth>=0.4.45
fitbit>=0.3.1
requests>=2.31.0
pyyaml>=6.0.1
fitparse>=1.2.0
garmin-fit-sdk>=21.141.0
schedule>=1.2.1
python-dateutil>=2.8.2
pytz
